Refactor module
Separate data, demo from views
Separate view, data and python scripts by model
Use CamelCase for class names

Nothing was changed to the data or code, just a refactoring
