--- mgmtsystem_nonconformity/board_mgmtsystem_nonconformity.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-
-        <record id="action_mgmtsystem_nonconformity" model="ir.actions.act_window">
-            <field name="res_model">mgmtsystem.nonconformity</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">tree</field>
-            <field name="view_id" ref="view_mgmtsystem_nonconformity_tree"/>
-            <field name="domain">[('state','=','open')]</field>
-        </record>
-
-        <record id="board_mgmtsystem_nonconformity_form" model="ir.ui.view">
-            <field name="name">board.mgmtsystem.nonconformity.form</field>
-            <field name="model">board.board</field>
-            <field name="inherit_id" ref="mgmtsystem.board_mgmtsystem_form"/>
-            <field name="arch" type="xml">
-                <xpath expr="/form/hpaned/child1" position="inside">
-		    <action name="%(action_mgmtsystem_nonconformity)d" string="Open Nonconformities"/>
-                </xpath>
-            </field>
-        </record>
-        
-    </data>
-</openerp>
