+++ mgmtsystem_nonconformity/data/mgmtsystem_nonconformity_severity.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data noupdate="1">
+
+        <!-- Severity: Critical, Major, Minor, Invalid, ... -->
+
+        <record id="nonconformity_severity_maj" model="mgmtsystem.nonconformity.severity">
+            <field name="name">Major</field>
+            <field name="active" eval="True"/>
+        </record>
+
+        <record id="nonconformity_severity_min" model="mgmtsystem.nonconformity.severity">
+            <field name="name">Minor</field>
+            <field name="active" eval="True"/>
+        </record>
+
+        <record id="nonconformity_severity_no" model="mgmtsystem.nonconformity.severity">
+            <field name="name">Unfounded</field>
+            <field name="active" eval="True"/>
+        </record>
+
+    </data>
+</openerp>
