+++ mgmtsystem_nonconformity/demo/mgmtsystem_nonconformity.xml
@@ -2,19 +2,6 @@
 <openerp>
     <data noupdate="1">
 
-        <!-- demo data mgmtsystem.nonconformity.origin -->
-        <record id="demo_origin" model="mgmtsystem.nonconformity.origin">
-            <field name="name">Internal Audit</field>
-            <field name="description">The nonconformity was discovered during an internal audit.</field>
-        </record>
-
-        <!-- demo data mgmtsystem.nonconformity.cause -->
-        <record id="demo_cause" model="mgmtsystem.nonconformity.cause">
-            <field name="name">Lack of expertise</field>
-            <field name="description">The cause of the nonconformity is a lack of expertise of the person.</field>
-        </record>
-
-        <!-- demo data mgmtsystem.nonconformity -->
         <record id="demo_nonconformity" model="mgmtsystem.nonconformity">
             <field name="partner_id" ref="base.res_partner_3"/>
             <field name="date">2010-12-03</field>
@@ -25,7 +12,7 @@
             <field name="description">The procedure has changed with no comments on the revision.</field>
             <field name="analysis">Administrator didn't know he had to add a comment when changing the procedure.</field>
             <field name="origin_ids" eval="[(6,0,[ref('demo_origin')])]"/>
-	    <field name="procedure_ids" eval="[(6,0,[ref('document_page_procedure.document_page_procedure')])]"/>
+            <field name="procedure_ids" eval="[(6,0,[ref('document_page_procedure.document_page_procedure')])]"/>
             <field name="cause_ids" eval="[(6,0,[ref('demo_cause')])]"/>
             <field name="immediate_action_id" ref="mgmtsystem_action.demo_immediate"/>
             <field name="corrective_action_id" ref="mgmtsystem_action.demo_corrective"/>
