+++ mgmtsystem_nonconformity/demo/mgmtsystem_nonconformity_cause.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data noupdate="1">
+
+        <record id="demo_cause" model="mgmtsystem.nonconformity.cause">
+            <field name="name">Lack of expertise</field>
+            <field name="description">The cause of the nonconformity is a lack of expertise of the person.</field>
+        </record>
+
+    </data>
+</openerp>
