+++ mgmtsystem_nonconformity/demo/mgmtsystem_nonconformity_origin.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data noupdate="1">
+
+        <record id="demo_origin" model="mgmtsystem.nonconformity.origin">
+            <field name="name">Internal Audit</field>
+            <field name="description">The nonconformity was discovered during an internal audit.</field>
+        </record>
+
+    </data>
+</openerp>
