+++ mgmtsystem_nonconformity/model/__init__.py
@@ -1,0 +1,9 @@
+# -*- encoding: utf-8 -*-
+
+from . import (
+    mgmtsystem_nonconformity_origin,
+    mgmtsystem_nonconformity_severity,
+    mgmtsystem_nonconformity_cause,
+    mgmtsystem_nonconformity,
+    mgmtsystem_action,
+)
